Added Arrow.html
Next project started, Arrow. will be shooting things and hopefully have a better gravity/angle system for projectiles.

Current issue is calculating the angle of mouse from start. I was thinking to have just the left side be colored(to let users know where they can click, or create a strip to click in on the mid-left side of screen), and read the Y value of the cursor, and use that based off of a fixed point using idk an equation lol.

Whatever value I have outputting, be it in radians or degrees (either are easily converted to each other smh) I still need to change the Gravity value appropriately, or add new variables as necessary.